Fix code style details, add HTML description
