+++ modules/GravityPlugin/src/main/java/org/gephi/plugins/layout/forceAtlas2/multigravity/Operation.java
@@ -39,23 +39,13 @@ Contributor(s):
 
 Portions Copyrighted 2011 Gephi Consortium.
  */
-package org.gephi.plugins.layout.forceAtlas2Custom;
-
-import org.gephi.graph.spi.LayoutData;
+package org.gephi.plugins.layout.forceAtlas2.multigravity;
 
 /**
- * Data stored in Nodes and used by ForceAtlas2
+ *
  * @author Mathieu Jacomy
  */
-public class ForceAtlas2LayoutData implements LayoutData {
-    //Data
+public abstract class Operation {
 
-    public double dx = 0;
-    public double dy = 0;
-    public double old_dx = 0;
-    public double old_dy = 0;
-    public double mass = 1;
-    // Gravity source
-    public double gravity_x = 0;
-    public double gravity_y = 0;
+    public abstract void execute();
 }
