+++ modules/GravityPlugin/src/main/java/org/gephi/plugins/layout/forceAtlas2/multigravity/OperationNodeNodeAttract.java
@@ -39,29 +39,31 @@ Contributor(s):
 
 Portions Copyrighted 2011 Gephi Consortium.
  */
-package org.gephi.plugins.layout.forceAtlas2Custom;
+package org.gephi.plugins.layout.forceAtlas2.multigravity;
 
 import org.gephi.graph.api.Node;
-import org.gephi.plugins.layout.forceAtlas2Custom.ForceFactory.RepulsionForce;
+import org.gephi.plugins.layout.forceAtlas2.multigravity.ForceFactory.AttractionForce;
 
 /**
  *
  * @author Mathieu Jacomy
  */
-public class OperationNodeNodeRepulse extends Operation {
+public class OperationNodeNodeAttract extends Operation {
 
     private final Node n1;
     private final Node n2;
-    private final RepulsionForce f;
+    private final AttractionForce f;
+    private final double coefficient;
 
-    public OperationNodeNodeRepulse(Node n1, Node n2, RepulsionForce f) {
+    public OperationNodeNodeAttract(Node n1, Node n2, AttractionForce f, double coefficient) {
         this.n1 = n1;
         this.n2 = n2;
         this.f = f;
+        this.coefficient = coefficient;
     }
 
     @Override
     public void execute() {
-        f.apply(n1, n2);
+        f.apply(n1, n2, coefficient);
     }
 }
