+++ modules/GravityPlugin/src/main/resources/org/gephi/plugins/layout/forceAtlas2/multigravity/Bundle.properties
@@ -1,5 +1,8 @@
+OpenIDE-Module-Long-Description=MultiGravity ForceAtlas 2. Nodes are expected to have an attribute which is "gravity_x" and "gravity_y" which need to be numbers, that give approximate locations of sources of gravity. The distance between the sources of gravity can be modified using the parameters for the layout. The main benefit of this algorithm is that graphs can be coerced into certain shapes based on some data related to the data set.
+
+
 ForceAtlas2.name=MultiGravity ForceAtlas 2
-ForceAtlas2.description=Quality layout: a linear-attraction linear-repulsion model with few approximations (BarnesHut). Speed automatically computed.
+ForceAtlas2.description=Quality layout: a linear-attraction linear-repulsion model with few approximations (BarnesHut). Speed automatically computed. Nodes are expected to have an attribute which is "gravity_x" and "gravity_y" which need to be numbers.
 
 ForceAtlas2.tuning=Tuning
 ForceAtlas2.behavior=Behavior Alternatives
